requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.24.3
scipy==1.11.2
sklearn
